Simplify types
