+++ crypto/Cargo.toml
@@ -35,6 +35,7 @@ serde_json = "1.0.85"
 thiserror = "1.0.34"
 tracing = "0.1.36"
 zeroize = "1.5.7"
+hex-literal = "0.3.4"
 
 [dev-dependencies]
 proptest = "1.0.0"
